requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
github3.py==0.9.5
gnureadline==6.3.3
ipython==4.2.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pexpect==4.1.0
pickleshare==0.7.2
prettytable==0.7.2
ptyprocess==0.5.1
PyYAML==3.11
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
uritemplate.py==0.3.0
